feat(vs2022_third): C4244 fix all conversions in RecipeDistanceStageProgramView.cpp — 1 file
Fix 15 C4244 warnings:
- Line 1501: Add casts for CRect constructor (4 conversions)
- Lines 1754-1756: Add casts for double→LONG assignments (3 conversions)
- Lines 2195, 2198, 2237, 2270: Add casts for MoveTo/LineTo arguments (8 conversions)

No behavior changes, only explicit type conversions.

+++ SRC/NanoSpec/RecipeDistanceStageProgramView.cpp
@@ -1498,7 +1498,7 @@ void CRecipeDistanceStageProgramView::OnPaint()
 		m_dDrawEndStagePosY = StageRectBottom;
 	}
 
-	pDC->Rectangle(CRect(m_dDrawStartStagePosX, m_dDrawStartStagePosY, m_dDrawEndStagePosX, m_dDrawEndStagePosY));
+	pDC->Rectangle(CRect(static_cast<int>(m_dDrawStartStagePosX), static_cast<int>(m_dDrawStartStagePosY), static_cast<int>(m_dDrawEndStagePosX), static_cast<int>(m_dDrawEndStagePosY)));
 	pDC->SelectObject(m_oldPen);
 	pDC->SelectObject(m_oldBrush);
